requirements.txt
tensorflow-cpu
pandas
Flask
sklearn
numpy
h5py==2.10.0
gunicorn